做了修改速度时的日志显示，验证通过，之前的XY 方向有问题

+++ src/serial_demo/src/myData.h
@@ -22,7 +22,7 @@ typedef struct _Motor_Parameter_
 {
   float target;
   uart_INT preloader;
-  uint8_t direction_Target;
+  int direction_Target;
   uart_INT encoder;
-  uint8_t direction_Now;
+  int direction_Now;
 } Motor_Parameter;
